0x01-python_async_function

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -4,7 +4,7 @@ import random
 import asyncio
 
 
-async def wait(delay_max: float = 10) -> float:
+async def wait(delay_max: int = 10) -> float:
     """Wait for a random time between 0 and delay_max seconds.
 
     Args:
